Archive 25/requirements.txt
pandas>=2.1.0
pydantic>=2.5.0
pyarrow>=14.0.0

loguru>=0.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Azure SDKs (replaces boto3 / AWS)
azure-storage-blob>=12.19.0
azure-storage-file-datalake>=12.15.0
azure-identity>=1.15.0

tenacity>=8.0.0
openpyxl>=3.1.0
mcp>=1.0.0
langgraph>=0.2.0
langfuse

boto3